requirements.txt
pdf2image
pdfminer.six
pyocr